% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHelp.R
\name{getHelp}
\alias{getHelp}
\title{getHelp \lifecycle{soft-deprecated}}
\usage{
getHelp(fn, pkg)
}
\arguments{
\item{fn}{character of function name}

\item{pkg}{name of package that contains fn}
}
\value{
chracter
}
\description{
return rd help file as html
}
