% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_team_logo-data.R
\docType{data}
\name{nba_team_logo}
\alias{nba_team_logo}
\title{nba_team_logo}
\format{A data frame with 30 rows and 2 variables:
\describe{
  \item{\code{team}}{character, Team Name}
  \item{\code{uri}}{character, Team logo URI} 
}}
\usage{
nba_team_logo
}
\description{
NBA team logo URI
}
\details{
Logos harvested from 
 https://i.cdn.turner.com/nba/nba/.element/img/4.0/global/logos/512x512/bg.white/svg/{team_short_name}.svg
}
\keyword{datasets}
