% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{\%synch\%}
\alias{\%synch\%}
\alias{\%stack\%}
\title{Combine slickR carousels}
\usage{
e1 \%synch\% e2

e1 \%stack\% e2
}
\arguments{
\item{e1}{slickR htmlwidget to place on top}

\item{e2}{slickR htmlwidget to place on bottom}
}
\value{
htmlwidget
}
\description{
Synchronize or Stack carousels
}
\details{
DETAILS
}
\examples{

if(interactive()){

# working with multiple slicks

sets <- split(
  sample(nba_team_logo$uri, size = 28, replace = FALSE),
  rep(c(1,2,3,4),each=7)
)

slicks <- lapply(sets,FUN = function(x,...){
  slickR(obj = x,...)
},height = 100)

# independent slicks

Reduce(`\%stack\%`,slicks)

# 1,2 synch stacked on 3,4 synch

 (slicks[[1]] \%synch\% slicks[[2]]) \%stack\% (slicks[[3]]\%synch\%slicks[[4]])
 
}
}
