% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dom_utils.R
\name{slick_list}
\alias{slick_list}
\title{List of DOM elements}
\usage{
slick_list(...)
}
\arguments{
\item{\dots}{Elements to convert}
}
\value{
tagList
}
\description{
Convert a list of \link[htmltools:builder]{tags} into a list
that slickR can render
}
\examples{

img1 <- htmltools::img(src = nba_player_logo$uri[1])
img2 <- htmltools::img(src = nba_player_logo$uri[2])

slick_list(img1,img2)

}
\seealso{
\link[htmltools:tagList]{tagList}, \link[htmltools:builder]{tags}

Other html: 
\code{\link{slick_div}()}
}
\concept{html}
