% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slick_magick.R
\name{slick_magick}
\alias{slick_magick}
\title{Convert PDF into Images}
\usage{
slick_magick(obj, dest = file.path(tempdir(), "slick", "pdf"),
  img_format = "png")
}
\arguments{
\item{obj}{character, path to PDF document. Can be local or URL address.}

\item{dest}{character, local directory to write images to,
Default: file.path(tempdir(), "slick", "pdf")}

\item{img_format}{character, image format to convert to, Default: 'png'}
}
\value{
character vector of files in destination directory
}
\description{
Convert PDF document into images to pass into slickR
}
\examples{
if(interactive()){
  pdf_file <- system.file('examples/slickR.pdf',package = 'slickR')
  imgs <- slick_magick(pdf_file)
  slickR(imgs,height = 500)
 }
}
