% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slickpdf.R
\name{slickpdf}
\alias{slickpdf}
\title{convert pdf to slick}
\usage{
slickpdf(obj, img_format = "png", ...)
}
\arguments{
\item{obj}{path to pdf(s)}

\item{img_format}{format of image to convert to, Default: 'png'}

\item{...}{arguments passed to \code{\link[slickR]{slickR}}}
}
\value{
slick
}
\description{
convert multiple pdfs or a multipage pdf to a slick carousel
}
\details{
\href{https://github.com/ropensci/pdftools}{pdftools} must be installed for function to work
}
\examples{
if(interactive()){
slickpdf('https://cran.r-project.org/web/packages/slickR/slickR.pdf')

slickpdf('https://cran.r-project.org/web/packages/slickR/slickR.pdf',imgFormat = 'svg')
}
 
 
}
\seealso{
\code{\link[magick]{image_read_pdf}}
}
