% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{slickpdf}
\alias{slickpdf}
\title{slickpdf \lifecycle{deprecated}}
\usage{
slickpdf(obj, obj2 = NULL, synch = TRUE, img_format = "png", ...)
}
\arguments{
\item{obj}{character, path to pdf(s)}

\item{obj2}{character, path to pdf(s) to compare to obj, Default: NULL}

\item{synch}{boolean, if TRUE then obj and obj2 images are synchronized, Default: TRUE}

\item{img_format}{format of image to convert to, Default: 'png'}

\item{...}{arguments passed to \code{\link[slickR]{slickR}}}
}
\value{
slick
}
\description{
convert multiple pdfs or a multipage pdf to a slick carousel
}
\details{
This function has been replaced with the example in the PDF vignette.
}
