% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slickR.R
\name{slickR}
\alias{slickR}
\title{slick.js image carousel htmlwidget}
\usage{
slickR(images, ..., width = NULL, height = NULL, elementId = NULL)
}
\description{
use slick.js library in R
}
\examples{
require(rvest)
a=c("ATL","BKN","BOS","CHA","CHI","CLE","DAL","DEN","DET","GSW",
"HOU","IND","LAC","LAL","MEM","MIA","MIL","MIN","NOP","NYK",
"OKC","ORL","PHI","PHX","POR","SAC","SAS","TOR","UTA","WAS")
x=sprintf("https://i.cdn.turner.com/nba/nba/.element/img/4.0/global/logos/512x512/bg.white/svg/\%s.svg",a)
slickR(images=x)
slickR(images = x,divName=c('slider-for','slider-nav'),imgWidth=c('50\%','20\%'))

}

