% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slickR.R
\name{slickR}
\alias{slickR}
\title{slick.js image carousel htmlwidget}
\usage{
slickR(images, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{images}{character, vector of path or url to images}

\item{...}{arguments to pass to slickjs, see details}

\item{width}{character, width of htmlwidget}

\item{height}{character, height of htmlwidget}

\item{elementId}{character, id tag of htmlwidget}
}
\description{
use slick.js library in R
}
\details{
slick.js \url{http://kenwheeler.github.io/slick/} is an image carousel javascript library.
to create more than one carousel input a a character vector `divName` and 
to control the width of images of the carousel use `imgWidth` (see example 2).
}
\examples{
require(rvest)
a=c("ATL","BKN","BOS","CHA","CHI","CLE","DAL","DEN","DET","GSW",
"HOU","IND","LAC","LAL","MEM","MIA","MIL","MIN","NOP","NYK",
"OKC","ORL","PHI","PHX","POR","SAC","SAS","TOR","UTA","WAS")
x=sprintf("https://i.cdn.turner.com/nba/nba/.element/img/4.0/global/logos/512x512/bg.white/svg/\%s.svg",a)
slickR(images=x)
slickR(images = x,divName=c('slider-for','slider-nav'),imgWidth=c('50\%','20\%'))

}

