% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slickR.R
\name{slickR}
\alias{slickR}
\title{slick.js image carousel htmlwidget}
\usage{
slickR(obj, slideId = "baseDiv", objLinks = NULL, slideType = "img",
  padding = 1, dotObj = NULL, width = "95\%", height = NULL,
  elementId = NULL)
}
\arguments{
\item{obj}{character, vector of path or url to images}

\item{slideId}{character, id of slide, Default: 'baseDiv'}

\item{objLinks}{character, links to attach to images in slide, Default: NULL}

\item{slideType}{character, type of object to put in slide, Default: 'img'}

\item{padding}{numeric, percent of width between each image in the carousel 
for each slider, Default: 1}

\item{dotObj}{character, urls or images to replace dots with (see details), 
Default: NULL}

\item{width}{character, width of htmlwidget, Default: '95\%'}

\item{height}{character, height of htmlwidget, Default: NULL}

\item{elementId}{character, id tag of htmlwidget, Default: NULL}
}
\description{
use slick.js library in R
}
\details{
slick.js \url{http://kenwheeler.github.io/slick/} is an image carousel javascript library. 

To find all the attributes that can be used please refer to the link. 

It is possible to stack slides through the `+` operator. 

It is possible to synchronize slides through the `*` operator. 

To replace the dots with icons use the dotObj argument to pass in the icon
images and in the slickOpts add a customPaging attribute with the appropriate JS(.) function call. 

The slideType accepts the type of html DOM you want to be in the slide, eg img, iframe.
}
\examples{


if(interactive()){

slick <- slickR(obj=nba_team_logo$uri)

slick

# add dots to the first slick and autoplay at 1 second a slide
 
slick + settings(dots = TRUE, autoplay = TRUE, autoplaySpeed = 1000)


# working with multiple slicks

sets <- split(
  sample(nba_team_logo$uri, size = 28, replace = FALSE),
  rep(c(1,2,3,4),each=7)
)

slicks <- lapply(sets,FUN = function(x,...){
  slickR(obj = x,...)
},height = 100)

# independent slicks

Reduce(`\%stack\%`,slicks)

# 1,2 synch stacked on 3,4 synch

 (slicks[[1]] \%synch\% slicks[[2]]) \%stack\% (slicks[[3]]\%synch\%slicks[[4]])

}
}
